feat: folder structure and return jwt token
