+++ src/routers/main_routes.py
@@ -1,0 +1,5 @@
+from config.app import app
+
+@app.get("/")
+async def root():
+    return {"message": "Hello World"}
